Revert "Added owls to questions, going to try and drag them around"
